can keep escape characters with noEscape option
