+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "@mnbroatch/bondage",
-  "version": "3.0.42",
+  "version": "3.1.0",
   "description": "Yarn parser for Javascript",
   "main": "dist/bondage.js",
   "scripts": {
     "test": "jest --verbose false",
     "test:coverage": "jest --coverage --verbose false",
     "lint": "eslint .",
-    "build": "jison-es src/parser/grammar.json -o src/parser/compiledParser.js && webpack",
+    "build": "jison-es src/parser/grammar.json -o src/parser/compiled-parser.js && webpack",
     "prepublish": "npm run build"
   },
   "repository": {
